requirements.txt
discord.py
python-dotenv
apscheduler
httpx
selectolax